+++ activerecord/lib/active_record/connection_adapters/mysql_adapter.rb
@@ -441,46 +441,5 @@ module ActiveRecord
         @version ||= @connection.server_info.scan(/^(\d+)\.(\d+)\.(\d+)/).flatten.map { |v| v.to_i }
       end
     end
-
-    module ClassColumn
-      def class_column( symbol, prefix )
-        klass = symbol.to_s.classify.constantize
-        #AR models do respond to a different columns method, but we only want to handle non-persistent classes
-        if klass.respond_to?(:columns) && !klass.ancestors.include?(ActiveRecord::Base)
-          klass.columns(prefix).each do |col_def|
-            column *col_def
-          end
-          return true
-        end
-      end
-    end
-
-    class Table
-      include ClassColumn
-
-      def change_column_null( column_name, null, default = nil )
-        @base.change_column_null(@table_name, column_name, null, default)
-      end
-
-      def method_missing(symbol, *args)
-        unless class_column(symbol, args[0])
-          super
-        end
-      end
-    end
-
-    class TableDefinition
-      include ClassColumn
-
-      def method_missing(symbol, *args)
-        if symbol.to_s == 'xml'
-          return xml_column_fallback(args)
-        end
-        unless class_column(symbol, args[0])
-          super
-        end
-      end
-    end
-
   end
 end
